Add hooks for l10n_ch_sepa (pain.001.001.03.ch.02)
Move code of account_banking_pain_base/models/banking_export_pain.py to account_banking_pain_base/models/account_payment_order.py
Allow cancel when payment order is in "File Generated" state
Add postal address in party block
