+++ account_banking_pain_base/models/__init__.py
@@ -7,5 +7,4 @@ from . import account_payment_order
 from . import bank_payment_line
 from . import account_payment_mode
 from . import res_company
-from . import banking_export_pain
 from . import account_payment_method
